\name{sheldus_assemble_data}
\alias{sheldus_assemble_data}
\title{Read pre-processed SHELDUS data and assemble variables into one data frame}
\usage{
  sheldus_assemble_data(years, hazards, year_vec,
    hazard_names)
}
\arguments{
  \item{years}{vector of years for which data is desired}

  \item{hazards}{vector of hazards names for which data is
  desired}

  \item{year_vec}{years for which SHELDUS database contains
  data}

  \item{hazard_names}{names of hazards from SHELDUS}
}
\description{
  Read pre-processed SHELDUS data and assemble variables
  into one data frame
}
\keyword{internal}

